+++ user/usys.pl
@@ -4,7 +4,7 @@
 
 print "# generated by usys.pl - do not edit\n";
 
-print "#include \"kernel/include/syscall.h\"\n";
+print "#include \"kernel/include/sys/syscall.h\"\n";
 
 sub entry {
     my $name = shift;
